requirements.txt
pandas>=2.1
numpy>=1.26
pydantic>=2.6
python-dotenv>=1.0
httpx>=0.27
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.8
scipy>=1.11
pytest>=8.0
